ML/agents/langchain/tools/requirements.txt
langchain
langchain-core
python-dotenv
openai

